# openapi.yaml
openapi: 3.1.0
info:
  title: Traccar
  description: Traccar GPS tracking server API documentation. To use the API you need to have a server instance. For testing purposes you can use one of free [demo servers](https://www.traccar.org/demo-server/). For production use you can install your own server or get a [subscription service](https://www.traccar.org/product/tracking-server/).
  contact:
    name: Traccar Support
    url: https://www.traccar.org/
    email: support@traccar.org
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: '6.14.5'
servers:
  - url: https://demo.traccar.org/api
    description: Demo Server 1
  - url: https://demo2.traccar.org/api
    description: Demo Server 2
  - url: https://demo3.traccar.org/api
    description: Demo Server 3
  - url: https://demo4.traccar.org/api
    description: Demo Server 4
  - url: https://server.traccar.org/api
    description: Subscription Server
  - url: http://{host}:{port}/api
    description: Other Server
    variables:
      host:
        default: localhost
      port:
        enum:
          - '8082'
          - '80'
        default: '8082'
security:
  - BasicAuth: []
  - ApiKey: []
...
paths:
...
  /permissions:
    get:
      operationId: getPermissions
      summary: Fetch permission links between Objects
      description: >-
        Provide exactly two `*Id` query parameters matching the Permission body
        shape (e.g. `reportId` and `deviceId`). Use `0` on a side to mean
        "any", e.g. `?reportId=5&deviceId=0` lists all devices linked to
        report 5. At least one side must be non-zero.
      tags:
        - Permissions
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Permission'
        '400':
          description: Bad request
          content: {}
    post:
      operationId: postPermissions
      summary: Link an Object to another Object
      tags:
        - Permissions
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Permission'
        required: true
      responses:
        '204':
          description: No Content
          content: {}
        '400':
          description: No permission
          content: {}
    delete:
      operationId: deletePermissions
      summary: Unlink an Object from another Object
      tags:
        - Permissions
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Permission'
        required: true
      responses:
        '204':
          description: No Content
          content: {}
        '401':
          description: Unauthorized
          content: {}
  /permissions/bulk:
    post:
      operationId: postPermissionsBulk
      summary: Link multiple Objects in a single request
      tags:
        - Permissions
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Permission'
        required: true
      responses:
        '204':
          description: No Content
          content: {}
        '400':
          description: Bad request
          content: {}
    delete:
      operationId: deletePermissionsBulk
      summary: Unlink multiple Objects in a single request
      tags:
        - Permissions
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Permission'
        required: true
      responses:
        '204':
          description: No Content
          content: {}
        '400':
          description: Bad request
          content: {}
...
  /server:
    get:
      operationId: getServer
      summary: Fetch Server information
      tags:
        - Server
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Server'
        '400':
          description: Bad request
          content: {}
    put:
      operationId: putServer
      summary: Update Server information
      tags:
        - Server
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Server'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Server'
        '401':
          description: Unauthorized
          content: {}
        '400':
          description: Bad request
          content: {}
...
  /session:
    get:
      operationId: getSession
      summary: Fetch Session information
      tags:
        - Session
      security: []
      parameters:
        - name: token
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Not found
          content: {}
    post:
      operationId: postSession
      summary: Create a new Session
      tags:
        - Session
      security: []
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                password:
                  type: string
                  format: password
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content: {}
    delete:
      operationId: deleteSession
      summary: Close the Session
      tags:
        - Session
      responses:
        '204':
          description: No Content
          content: {}
        '401':
          description: Unauthorized
          content: {}
...
  /users:
    get:
      operationId: getUsers
      summary: Fetch a list of Users
      tags:
        - Users
      parameters:
        - name: userId
          in: query
          description: Can only be used by admin or manager users
          schema:
            type: string
        - name: limit
          in: query
          description: Limit the number of returned results
          schema:
            type: integer
        - name: offset
          in: query
          description: Offset for pagination
          schema:
            type: integer
        - name: keyword
          in: query
          description: Search keyword filter (searches name, email)
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '400':
          description: No Permission
          content: {}
    post:
      operationId: postUsers
      summary: Create a User
      tags:
        - Users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content: {}
        '400':
          description: Bad request
          content: {}
  /users/totp:
    post:
      operationId: postUsersTotp
      summary: Generate TOTP secret
      tags:
        - Users
      security: []
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
        '400':
          description: Bad request
          content: {}
  /users/{id}:
    get:
      operationId: getUsersId
      summary: Fetch a User
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content: {}
        '404':
          description: Not found
          content: {}
    put:
      operationId: putUsersId
      summary: Update a User
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content: {}
        '400':
          description: Bad request
          content: {}
    delete:
      operationId: deleteUsersId
      summary: Delete a User
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: No Content
          content: {}
        '401':
          description: Unauthorized
          content: {}
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique user identifier
        name:
          type: string
          description: User display name
        email:
          type: string
          description: Email address used for login and notifications
        phone:
          type: [string, 'null']
          description: Contact phone number for alerts
        readonly:
          type: boolean
          description: When true, the user cannot change settings
        administrator:
          type: boolean
          description: Grants full administrative privileges when enabled
        map:
          type: [string, 'null']
          description: Preferred default map layer for the user
        latitude:
          type: number
          description: Default map center latitude for this user
        longitude:
          type: number
          description: Default map center longitude for this user
        zoom:
          type: integer
          description: Default map zoom level on login
        password:
          type: string
          description: Password for user authentication
        coordinateFormat:
          type: [string, 'null']
          description: Preferred coordinate display format
        disabled:
          type: boolean
          description: Indicates whether the user account is disabled
        expirationTime:
          type: [string, 'null']
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          format: date-time
        deviceLimit:
          type: integer
          description: Maximum number of devices the user can manage
        userLimit:
          type: integer
          description: Maximum number of subordinate users
        deviceReadonly:
          type: boolean
          description: Restricts user from modifying device attributes
        limitCommands:
          type: boolean
          description: Prevents user from sending unsupported commands
        fixedEmail:
          type: boolean
          description: Locks the email field to avoid changes
        poiLayer:
          type: [string, 'null']
          description: External POI layer configured for the user
        attributes:
          type: object
          description: Additional custom user attributes
          properties: {}
    Server:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique server configuration identifier
        registration:
          type: boolean
          description: Whether new user registrations are allowed
        readonly:
          type: boolean
          description: When true only administrators can modify server-wide settings
        deviceReadonly:
          type: boolean
          description: Disallow device attribute changes for non-admins
        limitCommands:
          type: boolean
          description: Restrict command execution to supported protocol commands
        map:
          type: string
          description: Default map layer identifier
        bingKey:
          type: string
          description: Bing Maps API key used when Bing is selected as a provider
        mapUrl:
          type: string
          description: Custom tile server URL template if configured
        poiLayer:
          type: string
          description: External point-of-interest layer configuration
        announcement:
          type: string
          description: Message displayed to all users in the web application
        latitude:
          type: number
          description: Default map center latitude
        longitude:
          type: number
          description: Default map center longitude
        zoom:
          type: integer
          description: Default map zoom level
        version:
          type: string
          description: Traccar server version string
        forceSettings:
          type: boolean
          description: Forces users to use the server-wide settings instead of their own
        coordinateFormat:
          type: string
          description: Default coordinate format for displaying positions
        openIdEnabled:
          type: boolean
          description: Indicates whether OpenID authentication is available
        openIdForce:
          type: boolean
          description: Require OpenID authentication for all users when enabled
        attributes:
          type: object
          description: Additional server-level configuration values
          properties: {}
...
    Permission:
      type: object
      properties:
        userId:
          type: integer
          format: int64
          description: User id, can be only first parameter
        deviceId:
          type: integer
          format: int64
          description: >-
            Device id, can be first parameter or second only in combination with
            userId
        groupId:
          type: integer
          format: int64
          description: >-
            Group id, can be first parameter or second only in combination with
            userId
        geofenceId:
          type: integer
          format: int64
          description: Geofence id, can be second parameter only
        notificationId:
          type: integer
          format: int64
          description: Notification id, can be second parameter only
        calendarId:
          type: integer
          format: int64
          description: >-
            Calendar id, can be second parameter only and only in combination
            with userId
        attributeId:
          type: integer
          format: int64
          description: Computed attribute id, can be second parameter only
        driverId:
          type: integer
          format: int64
          description: Driver id, can be second parameter only
        managedUserId:
          type: integer
          format: int64
          description: >-
            User id, can be second parameter only and only in combination with
            userId
        commandId:
          type: integer
          format: int64
          description: Saved command id, can be second parameter only
      description: >-
        This is a permission map that contain two object indexes. It is used to
        link/unlink objects. Order is important. Example: { deviceId:8,
        geofenceId: 16 }
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    ApiKey:
      type: http
      scheme: bearer